% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{download_res}
\alias{download_res}
\title{Downloads an open data record}
\usage{
download_res(resId)
}
\arguments{
\item{resId}{String of the resource id associated with the open data resource}
}
\value{
File path of the directory
}
\description{
This is a helper function to download and unzip an open data resource into a
temporary directory. It returns the path of the directory that the files
were downloaded into. The directory and downloaded data should be deleted
using unlink().
}
