% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{download_url}
\alias{download_url}
\title{Downloads an open data record}
\usage{
download_url(resUrl)
}
\arguments{
\item{resUrl}{String of url to download the resource}
}
\value{
File path of the directory
}
\description{
This is a helper function to download and unzip an open data resource into a
temporary directory. It returns the path of the directory that the files
were downloaded into. The directory and downloaded data should be deleted
using unlink().
}
\examples{
dataDir <- download_url("https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/110m/physical/ne_110m_ocean.zip")
list.files(dataDir)
}
