% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{get_esri_layer}
\alias{get_esri_layer}
\title{Read an esri layer}
\usage{
get_esri_layer(esriLayer, esriUrl, where, geometry)
}
\arguments{
\item{esriLayer}{String name of the layer to be read}

\item{esriUrl}{string url of the mapserver}

\item{where}{where clause to filter feature rows}

\item{geometry}{geometry_sf to filter feature rows}
}
\value{
sf object
}
\description{
This is a helper function to read in a geodatabase layer to an sf object.
}
