% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{get_gdb_layer}
\alias{get_gdb_layer}
\title{Read a geodatabase layer}
\usage{
get_gdb_layer(gdbLayer, gdbPath)
}
\arguments{
\item{gdbLayer}{String name of the layer to be read}

\item{gdbPath}{file path of the geodatabase}
}
\value{
sf object
}
\description{
This is a helper function to read in a geodatabase layer to an sf object.
}
