% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{get_star_tif}
\alias{get_star_tif}
\title{Read a Tif file}
\usage{
get_star_tif(tifPath, crs, crop_sf)
}
\arguments{
\item{tifPath}{File path of the tif file}

\item{crs}{Coordinate reference system used for output sf}

\item{crop_sf}{Sf object to crop the spatial data to}
}
\value{
stars object
}
\description{
This is a helper function to read in a tif file to stars object.
}
