% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{is_zip}
\alias{is_zip}
\title{Check if a file is a zip file}
\usage{
is_zip(filepath)
}
\arguments{
\item{filepath}{filepath of possible zip file}
}
\value{
boolean TRUE if file is zipped, else FALSE
}
\description{
Helper function that sees if a file can be unziped or not.
}
