% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{get_shp_res}
\alias{get_shp_res}
\title{Read an open data record}
\usage{
get_shp_res(resId, cropSf = NULL, crs = 4326)
}
\arguments{
\item{resId}{String of the resource id associated with the open data resource}

\item{cropSf}{Sf object to crop the spatial data too}

\item{crs}{Coordinate reference system used for output sf}
}
\value{
An sf object
}
\description{
These functions download and import open data spatial data based on a
resource id. The default data format for the output is an sf object, with
the option of cropping results to an input sf object.
}
\section{Functions}{
\itemize{
\item \code{get_shp_res}:
}}

\section{After function section}{

Maybe details here?
}

