% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{sf_clean}
\alias{sf_clean}
\title{Performs cleaning of an sf object}
\usage{
sf_clean(in_sf, crop_sf, crs)
}
\arguments{
\item{in_sf}{Sf object to be cleaned.}

\item{crop_sf}{Sf object to crop the spatial data to}

\item{crs}{Coordinate reference system used for output sf}
}
\value{
File path of the directory
}
\description{
This is a helper function to perform common cleaning operations on sf
objects. These include setting the geometry type to multipolygon if there are
multiple geometry types, running sf::st_make_valid, cropping to an sf if provided
and setting the crs.
}
